+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added the `messaging` package which adds support for Firebase Cloud Messaging.
 
 ## Changes
-* `Firebase` struct in the `database` package was renamed to `database`
+* Removed request timeout logic from `database` package.
+* `Firebase` struct in the `database` package was renamed to `Database`
 * The realtime database code has been moved from `github.com/zabawaba99/firego`
 to `github.com/zabawaba99/firego/database`